fix(ui): store floats for box

+++ invokeai/frontend/web/src/features/controlLayers/store/types.ts
@@ -38,10 +38,10 @@ export const zImageWithDims = z.object({
 export type ImageWithDims = z.infer<typeof zImageWithDims>;
 
 const zCropBox = z.object({
-  x: z.number().int().min(0),
-  y: z.number().int().min(0),
-  width: z.number().int().positive(),
-  height: z.number().int().positive(),
+  x: z.number().min(0),
+  y: z.number().min(0),
+  width: z.number().positive(),
+  height: z.number().positive(),
 });
 export const zCroppableImageWithDims = z.object({
   original: zImageWithDims,
